+++ src/main/java/nextstep/infrastructure/LoginMember.java
@@ -1,4 +1,4 @@
-package nextstep.auth;
+package nextstep.infrastructure;
 
 public class LoginMember {
     private Long id;
