+++ src/test/java/nextstep/auth/JwtTokenProviderTest.java
@@ -1,5 +1,6 @@
 package nextstep.auth;
 
+import nextstep.infrastructure.JwtTokenProvider;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
